chore: reformat

+++ src/core/dns/DNSClient.cpp
@@ -28,14 +28,13 @@ UdpDNSResponse *DNSClient::udpDns(const vector<string> &domains, const string &d
 }
 
 
-TcpDNSResponse *DNSClient::tcpDns(const string &domain, const string &dnsServer, uint16_t port, uint32_t timeout) {
+TcpDNSResponse *DNSClient::tcpDns(const string &domain, const string &dnsServer, uint16_t port, uint64_t timeout) {
     vector<string> domains;
     domains.emplace_back(domain);
     return tcpDns(domains, dnsServer, port, timeout);
 }
 
-TcpDNSResponse *
-DNSClient::tcpDns(const vector<string> &domains, const string &dnsServer, uint16_t port, uint32_t timeout) {
+TcpDNSResponse *DNSClient::tcpDns(const vector<string> &domains, const string &dnsServer, uint16_t port, uint64_t timeout) {
     return instance.queryTcp(domains, dnsServer, port, timeout);
 }
 
@@ -46,16 +45,15 @@ UdpDNSResponse *DNSClient::queryUdp(const vector<string> &domains, const string 
     deadline_timer timeout(ioContext);
     unsigned short qid = dnsRequest.dnsHeader->id;
     UdpDNSResponse *dnsResponse = nullptr;
-    std::future<size_t> future = socket.async_send_to(buffer(dnsRequest.data, dnsRequest.len),
-                                                      udp::endpoint(make_address_v4(dnsServer), port),
+    std::future<size_t> future = socket.async_send_to(buffer(dnsRequest.data, dnsRequest.len), udp::endpoint(make_address_v4(dnsServer), port),
                                                       boost::asio::use_future);
     future_status status = future.wait_for(std::chrono::milliseconds(1500));
     if (status != std::future_status::ready) {
         Logger::ERROR << dnsRequest.dnsHeader->id << "connect timeout!" << END;
     } else {
         dnsResponse = new UdpDNSResponse(1024);
-        std::future<size_t> receiveFuture = socket.async_receive_from(buffer(dnsResponse->data, sizeof(byte) * 1024),
-                                                                      serverEndpoint, boost::asio::use_future);
+        std::future<size_t> receiveFuture = socket.async_receive_from(buffer(dnsResponse->data, sizeof(byte) * 1024), serverEndpoint,
+                                                                      boost::asio::use_future);
         bool receiveStatus = receiveFuture.wait_for(std::chrono::milliseconds(5000)) == std::future_status::ready;
         if (!receiveStatus) {
             Logger::ERROR << dnsRequest.dnsHeader->id << "receive timeout!" << END;
@@ -97,8 +95,7 @@ DNSClient::DNSClient() {
     });
 }
 
-TcpDNSResponse *
-DNSClient::queryTcp(const vector<string> &domains, const string &dnsServer, uint16_t port, uint32_t timeout) {
+TcpDNSResponse *DNSClient::queryTcp(const vector<string> &domains, const string &dnsServer, uint16_t port, uint64_t timeout) {
     tcp::endpoint serverEndpoint(make_address_v4(dnsServer), port);
     TcpDnsRequest dnsRequest(domains);
     unsigned short qid = dnsRequest.dnsHeader->id;
@@ -110,64 +107,43 @@ DNSClient::queryTcp(const vector<string> &domains, const string &dnsServer, uint
     uint16_t length = 0;
     bool error = false;
     long begin = time::now();
-    auto connectFuture = socket.lowest_layer().async_connect(serverEndpoint,
-                                                             boost::asio::use_future([](boost::system::error_code ec) {
-                                                                 return true;
-                                                             }));
+    auto connectFuture = socket.lowest_layer().async_connect(serverEndpoint, boost::asio::use_future([](boost::system::error_code ec) {
+        return true;
+    }));
     future_status connectStatus = connectFuture.wait_for(std::chrono::milliseconds(timeout));
     if (connectStatus != std::future_status::ready) {
         Logger::ERROR << dnsRequest.dnsHeader->id << "connect timeout!" << END;
         error = true;
     } else {
-        timeout -= (
-
-                time::now()
-
-                - begin);
+        timeout -= (time::now() - begin);
         begin = time::now();
         auto value = connectFuture.get();
         auto shakeFuture = socket.async_handshake(boost::asio::ssl::stream_base::client, buffer(sslByteBuf, 10240),
-                                                  boost::asio::use_future(
-                                                          [](boost::system::error_code error, size_t asd) {
-                                                              return error;
-                                                          }));
+                                                  boost::asio::use_future([](boost::system::error_code error, size_t asd) {
+                                                      return error;
+                                                  }));
         if (shakeFuture.wait_for(std::chrono::milliseconds(timeout)) != std::future_status::ready) {
             Logger::ERROR << dnsRequest.dnsHeader->id << "ssl handshake timeout!" << END;
             error = true;
         } else {
-            timeout -= (
-
-                    time::now()
-
-                    - begin);
+            timeout -= (time::now() - begin);
             begin = time::now();
-            auto sendFuture = boost::asio::async_write(socket, buffer(dnsRequest.data, dnsRequest.len),
-                                                       boost::asio::use_future);
+            auto sendFuture = boost::asio::async_write(socket, buffer(dnsRequest.data, dnsRequest.len), boost::asio::use_future);
             if (sendFuture.wait_for(std::chrono::milliseconds(timeout)) != std::future_status::ready) {
                 Logger::ERROR << dnsRequest.dnsHeader->id << "send dns request timeout!" << END;
                 error = true;
             } else {
-                timeout -= (
-
-                        time::now()
-
-                        - begin);
+                timeout -= (time::now() - begin);
                 begin = time::now();
-                auto receiveLenFuture = boost::asio::async_read(socket, buffer(lengthBytes, 2),
-                                                                boost::asio::use_future);
+                auto receiveLenFuture = boost::asio::async_read(socket, buffer(lengthBytes, 2), boost::asio::use_future);
                 if (receiveLenFuture.wait_for(std::chrono::milliseconds(timeout)) != std::future_status::ready) {
                     Logger::ERROR << dnsRequest.dnsHeader->id << "receive dns response len timeout!" << END;
                     error = true;
                 } else {
-                    timeout -= (
-
-                            time::now()
-
-                            - begin);
+                    timeout -= (time::now() - begin);
                     begin = time::now();
                     st::utils::read(lengthBytes, length);
-                    auto receiveDataFuture = boost::asio::async_read(socket, buffer(dataBytes, length),
-                                                                     boost::asio::use_future);
+                    auto receiveDataFuture = boost::asio::async_read(socket, buffer(dataBytes, length), boost::asio::use_future);
                     if (receiveDataFuture.wait_for(std::chrono::milliseconds(timeout)) != std::future_status::ready) {
                         Logger::ERROR << dnsRequest.dnsHeader->id << "receive dns response data timeout!" << END;
                         error = true;
@@ -180,33 +156,11 @@ DNSClient::queryTcp(const vector<string> &domains, const string &dnsServer, uint
     }
 
     try {
-        socket.
-
-                lowest_layer()
-
-                .
-
-                        cancel();
-
-        socket.
-
-                lowest_layer()
-
-                .shutdown(boost::asio::socket_base::shutdown_type::shutdown_both);
-        socket.
-
-                lowest_layer()
-
-                .
-
-                        close();
-
+        socket.lowest_layer().cancel();
+        socket.lowest_layer().shutdown(boost::asio::socket_base::shutdown_type::shutdown_both);
+        socket.lowest_layer().close();
     } catch (boost::wrapexcept<boost::system::system_error> ex) {
-        Logger::ERROR << dnsRequest.dnsHeader->id << ex.
-
-                what()
-
-                      << END;
+        Logger::ERROR << dnsRequest.dnsHeader->id << ex.what() << END;
     }
 
     TcpDNSResponse *dnsResponse = nullptr;
@@ -215,11 +169,7 @@ DNSClient::queryTcp(const vector<string> &domains, const string &dnsServer, uint
         st::utils::copy(lengthBytes, dnsResponse->data, 0, 0, 2);
         st::utils::copy(dataBytes, dnsResponse->data + 2, 0, 0, length);
         dnsResponse->parse(2 + length);
-        if (!dnsResponse->
-
-                isValid()
-
-                ) {
+        if (!dnsResponse->isValid()) {
             Logger::ERROR << dnsRequest.dnsHeader->id << "receive unmarketable data" << END;
             error = true;
         } else {
@@ -229,8 +179,7 @@ DNSClient::queryTcp(const vector<string> &domains, const string &dnsServer, uint
                     error = true;
                 }
                 if (dnsResponse->udpDnsResponse->header->responseCode != 0) {
-                    Logger::ERROR << dnsRequest.dnsHeader->id << "receive error responseCode"
-                                  << dnsResponse->udpDnsResponse->header->responseCode << END;
+                    Logger::ERROR << dnsRequest.dnsHeader->id << "receive error responseCode" << dnsResponse->udpDnsResponse->header->responseCode << END;
                     error = true;
                 }
             }
